requirements.txt
alembic==1.0.5
atomicwrites==1.2.1
attrs==18.2.0
bcrypt==3.1.5
cffi==1.11.5
Click==7.0
configparser==3.5.0
Flask==1.0.2
Flask-Migrate==2.3.1
Flask-SQLAlchemy==2.3.2
funcsigs==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
more-itertools==5.0.0
pathlib2==2.3.3
pluggy==0.8.0
py==1.7.0
pycparser==2.19
PyJWT==1.7.1
pytest==4.0.2
pytest-ordering==0.6
python-dateutil==2.7.5
python-editor==1.0.3
scandir==1.9.0
six==1.12.0
SQLAlchemy==1.2.15
uuid==1.30
Werkzeug==0.14.1
